requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.5
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
markdown2==2.5.2
multidict==6.1.0
nodeenv==1.9.1
openai==1.59.8
packaging==24.2
pdfkit==1.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
virtualenv==20.29.1
wheel==0.45.1
yarl==1.18.3
